% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen4R_citation.R
\name{get_citation}
\alias{get_citation}
\title{get_citation}
\usage{
get_citation(
  doi,
  sandbox = FALSE,
  style = c("havard-cite-them-right", "apa", "modern-language-association", "vancouver",
    "chicago-fullnote-bibliography", "ieee"),
  logger = NULL
)
}
\arguments{
\item{doi}{a Zenodo DOI or concept DOI}

\item{sandbox}{Use the sandbox infrastructure. Default is \code{FALSE}}

\item{style}{the style character string among. Possible values "havard-cite-them-right", "apa",
"modern-language-association","vancouver","chicago-fullnote-bibliography", or "ieee"}

\item{logger}{a logger to print messages. The logger can be either NULL, 
"INFO" (with minimum logs), or "DEBUG" (for complete curl http calls logs)}
}
\value{
an object of class \code{data.frame} giving the record versions
including date, version number and version-specific DOI.

the record citation
}
\description{
\code{get_citation} allows to get the styled citation for Zenodo record, identified by 
its DOI or concept DOI, and the style name
}
\examples{
\dontrun{
 get_citation(doi = "10.5281/zenodo.2547036", style = "ieee")
}
 
}
