% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen4R_getter.R
\name{get_zenodo}
\alias{get_zenodo}
\title{get_zenodo}
\usage{
get_zenodo(doi, logger = NULL)
}
\arguments{
\item{doi}{a Zenodo DOI or concept DOI}

\item{logger}{a logger to print messages. The logger can be either NULL, 
"INFO" (with minimum logs), or "DEBUG" (for complete curl http calls logs)}
}
\value{
an object of class \code{data.frame} giving the record versions
including date, version number and version-specific DOI.

object of class \code{ZenodoRecord}
}
\description{
\code{get_zenodo} allows to get a Zenodo record, identified by 
its DOI or concept DOI.
}
\examples{
\dontrun{
 get_zenodo("10.5281/zenodo.2547036")
}
                
}
