% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen4R_downloader.R
\name{human_filesize}
\alias{human_filesize}
\title{Human-readable binary file size}
\usage{
human_filesize(x)
}
\arguments{
\item{x}{A positive integer, i.e. the number of bytes (B).
Can be a vector of file sizes.}
}
\value{
A character vector.
}
\description{
Takes an integer (referring to number of bytes) and returns an optimally
human-readable
\href{https://en.wikipedia.org/wiki/Binary_prefix}{binary-prefixed}
byte size (KiB, MiB, GiB, TiB, PiB, EiB).
The function is vectorised.
}
\author{
Floris Vanderhaeghe, \email{floris.vanderhaeghe@inbo.be}
}
\keyword{internal}
