% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen4R_pat.R
\name{zenodo_pat}
\alias{zenodo_pat}
\title{zenodo_pat}
\usage{
zenodo_pat(quiet = TRUE)
}
\arguments{
\item{quiet}{Hide log message, default is \code{TRUE}}
}
\description{
Get Zenodo personal access token, looking in env var `ZENODO_PAT`
}
