% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen4R_licenses.R
\name{get_licenses}
\alias{get_licenses}
\title{get_licenses}
\usage{
get_licenses(pretty = TRUE)
}
\arguments{
\item{pretty}{output delivered as \code{data.frame}}
}
\value{
the licenses as \code{list} or \code{data.frame}
}
\description{
\code{get_licenses} allows to list all licenses supported by Zenodo.
}
\examples{
\dontrun{
 get_licenses(pretty = TRUE)
}
                
}
