% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen4R_versioning.R
\name{get_versions}
\alias{get_versions}
\title{get_versions}
\usage{
get_versions(doi, logger = NULL)
}
\arguments{
\item{doi}{a Zenodo DOI or concept DOI}

\item{logger}{a logger to print messages. The logger can be either NULL, 
"INFO" (with minimum logs), or "DEBUG" (for complete curl http calls logs)}
}
\value{
an object of class \code{data.frame} giving the record versions
including date, version number and version-specific DOI.
}
\description{
\code{get_versions} allows to execute a workflow
}
\examples{
\dontrun{
get_versions("10.5281/zenodo.2547036")
}
                
}
