% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen4R.R
\docType{package}
\name{zen4R}
\alias{zen4R}
\alias{zen4R-package}
\title{Interface to 'Zenodo' REST API}
\description{
Provides an Interface to 'Zenodo' (<https://zenodo.org>) REST API, 
  including management of depositions, attribution of DOIs by 'Zenodo', 
  upload and download of files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eblondel/zen4R}
  \item Report bugs at \url{https://github.com/eblondel/zen4R/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
