% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen4R.R
\docType{package}
\name{zen4R}
\alias{zen4R}
\alias{zen4R-package}
\alias{zen4R-package}
\title{Interface to Zenodo REST API}
\description{
Provides an Interface to Zenodo REST API
}
\details{
\tabular{ll}{
   Package: \tab zen4R\cr
   Type: \tab Package\cr
   Version: \tab 0.1\cr
   Date: \tab 2019-01-21\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
