% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen4R.R
\docType{package}
\name{zen4R}
\alias{zen4R}
\alias{zen4R-package}
\title{Interface to 'Zenodo' REST API}
\description{
Provides an Interface to 'Zenodo' (<https://zenodo.org>) REST API, 
  including management of depositions, attribution of DOIs by 'Zenodo' and 
  upload of files.
}
\details{
\tabular{ll}{
   Package: \tab zen4R\cr
   Type: \tab Package\cr
   Version: \tab 0.5\cr
   Date: \tab 2021-09-14\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
